完成代码整理
