--- Sample/Godot/用Resources作为数据源/Data.gd
@@ -1,4 +1,0 @@
-class_name Data
-extends Resrouces
-
-export(Dictionary) var data
