--- Sample/Godot/用Resources作为数据源/Settings.gd
@@ -1,5 +1,0 @@
-
-class_name Settings
-extends Object
-
-const var Item = preload(f'res://Settings/道具/Item.tres')
