--- Sample/Godot/用Resources作为数据源/export.toml
@@ -1,6 +1,0 @@
-[settings]
-ignore_sheet_mark = "*"
-custom_generator = "CustomGenerator/generator.py"
-completed_hook = "CustomGenerator/completed_hook.py"
-input = "Excel"
-output = "Settings"
