--- Sample/Godot/直接用GDScript作为数据源/README.md
@@ -1,9 +1,0 @@
-# 如何在GD脚本中读表
-
-```python
-
-func _ready():
-
-    var data = Settings.Item.data()
-
-```
