--- Sample/Godot/直接用GDScript作为数据源/run.bat
@@ -1,1 +1,0 @@
-ee gen-all
