+++ src/excelexporter/cli.py
@@ -1,6 +1,8 @@
+import shutil
 import click
 import textwrap
 import os
+import pkg_resources
 from . import exporter
 
 
@@ -48,18 +50,17 @@ def create_generator(filename="custom_generator.py"):
     # import jinja
     # 然后使用jinja模板来生成你的文件
 
-    # 下面是示例，直接导出保存成json文件，你可以删掉下面代码用jinja模板工具
-    # 用模板生成你的代码文件，然后写入output路径保存成代码文件
-
-    def gen(data, output):
+    def generator(data:dict, CFG:dict) -> (str, str):
         # 在这里面写你的加工逻辑
-        import json
-        with open(output+".json",'w+', encoding='utf-8') as f:
-            json.dump(data, f, ensure_ascii=False, indent=4)
+        # data是excel sheet转成字典
+        # CFG字典内容跟export.toml是一致的，你可以通过CFG获取到input、output、project_root
+        # 返回值 （你生成的代码, 导出保存的后缀名（例如gd、tres））
 
-    gen(data,output)
+        code = ""
+        return code, 'gd'
+
     """
-    with open(filename + ".py", "w+", encoding="utf-8") as f:
+    with open(filename + ".py", "w", encoding="utf-8") as f:
         f.write(textwrap.dedent(code))
 
 
@@ -76,7 +77,7 @@ def create_completed_hook():
     # 最典型的用法是用来搜集所有配置表生成读表util:
 
     # class_name Settiings
-    # Settings.gd
+    # settings.gd
     # const var Item = preload("res://Data/Item/Item.gd")
     # ...
 
@@ -85,9 +86,17 @@ def create_completed_hook():
 
     # todo:具体要怎么实现看自己需求
 
+    def completed(CFG:dict):
+        # CFG字典内容跟export.toml是一致的，你可以通过CFG获取到input、output、project_root
+        # 返回值 你生成的代码
+        # todo: 你要在导表结束后做什么
+        
+        code = ""
+        return code
+
     """
     code = textwrap.dedent(code)
-    with open("completed_hook.py", "w+") as f:
+    with open("completed_hook.py", "w", encoding='utf-8') as f:
         f.write(code)
 
 
@@ -97,8 +106,18 @@ def init():
     生成默认配置表项目
     """
     exporter.create_default_toml_config()
-    os.mkdir(exporter.CFG["settings"]["input"])
-    os.mkdir(exporter.CFG["settings"]["output"])
+    input_dir = exporter.CFG["settings"]["input"]
+    output_dir = exporter.CFG["settings"]["output"]
+
+    template_xlsx_path = pkg_resources.resource_filename(
+        __package__,
+        "template/示例.xlsx"
+    )
+
+    os.mkdir(input_dir)
+    os.mkdir(output_dir)
+
+    shutil.copy(template_xlsx_path, input_dir)
 
 
 @main.command()
